Move app deps to index file
