+++ chat-server/src/app.ts
@@ -6,15 +6,12 @@ import { llm } from "./services/llm";
 import {
   createMessage,
   logger,
-  MongoDB,
-  CORE_ENV_VARS,
-  assertEnvVars,
-  makeOpenAiEmbedFunc,
-  makeDatabaseConnection,
+  EmbeddedContentStore,
+  EmbedFunc,
 } from "chat-core";
-import { DataStreamerService } from "./services/dataStreamer";
+import { DataStreamerServiceInterface } from "./services/dataStreamer";
 import { ObjectId } from "mongodb";
-import { ConversationsService } from "./services/conversations";
+import { ConversationsServiceInterface } from "./services/conversations";
 
 // General error handler; called at usage of next() in routes
 const errorHandler: ErrorRequestHandler = (err, _req, res) => {
@@ -37,41 +34,17 @@ const reqHandler: RequestHandler = (req, _res, next) => {
   next();
 };
 
-export const makeApp = async (): Promise<
-  Express & { close(): Promise<void> }
-> => {
-  const {
-    MONGODB_CONNECTION_URI,
-    MONGODB_DATABASE_NAME,
-    VECTOR_SEARCH_INDEX_NAME,
-    OPENAI_ENDPOINT,
-    OPENAI_API_KEY,
-    OPENAI_EMBEDDING_DEPLOYMENT,
-    OPENAI_EMBEDDING_MODEL_VERSION,
-  } = assertEnvVars(CORE_ENV_VARS);
-
-  // Create instances of services
-  const mongodb = new MongoDB(
-    MONGODB_CONNECTION_URI,
-    MONGODB_DATABASE_NAME,
-    VECTOR_SEARCH_INDEX_NAME
-  );
-
-  const conversationsService = new ConversationsService(mongodb.db);
-  const dataStreamer = new DataStreamerService();
-
-  const store = await makeDatabaseConnection({
-    connectionUri: MONGODB_CONNECTION_URI,
-    databaseName: MONGODB_DATABASE_NAME,
-  });
-
-  const embed = makeOpenAiEmbedFunc({
-    apiKey: OPENAI_API_KEY,
-    apiVersion: OPENAI_EMBEDDING_MODEL_VERSION,
-    baseUrl: OPENAI_ENDPOINT,
-    deployment: OPENAI_EMBEDDING_DEPLOYMENT,
-  });
-
+export const makeApp = async ({
+  embed,
+  dataStreamer,
+  store,
+  conversationsService,
+}: {
+  embed: EmbedFunc;
+  store: EmbeddedContentStore;
+  dataStreamer: DataStreamerServiceInterface;
+  conversationsService: ConversationsServiceInterface;
+}): Promise<Express> => {
   const app = express();
   // TODO: consider only serving this from the staging env
   app.use(express.static("static"));
@@ -93,10 +66,5 @@ export const makeApp = async (): Promise<
   });
   app.use(errorHandler);
 
-  return Object.assign(app, {
-    async close() {
-      await mongodb.close();
-      await store.close();
-    },
-  });
+  return app;
 };
